[Improve][Connector-V2] Refactor file base
